1.1
php fixes

+++ php-mysql-cp/deluserid.php
@@ -1,7 +1,7 @@
 <?php
 include 'master/config.php'; 
 
-if(isset($_GET['id'])&&$_GET['id']!=1){
+if(isset($_GET['id'])&&$_GET['id']!=1&&$_SESSION["loggedin"]){
     $id = (int) $_GET['id'];
     if(!empty($_GET['id'])) {
         $delete = mysqli_query($conn,"DELETE FROM users WHERE id='$id'");
